Move DoFnInfo to SDK util
Previously, the DoFnInfo wrapped things just enough for Dataflow to execute a
DoFn without much context. The Java SDK harness has the same need, and relies
on DoFnInfo. Effectively, DoFnInfo is the UDF that the Java SDK harness
understands.
